플레이리스트 수정시 그룹아이디가 맞지 않을 경우 처리 되지 않도록 적용

+++ playlists/views.py
@@ -73,11 +73,12 @@ class PlayListViewSet(viewsets.ModelViewSet):
 
                         try:
                             obj = PlayList.objects.get(id=_id)
-                            if is_delete:
-                                obj.delete()
-                            else:
-                                obj.order = order
-                                obj.save()
+                            if obj.group_id == pk:
+                                if is_delete:
+                                    obj.delete()
+                                else:
+                                    obj.order = order
+                                    obj.save()
 
                         except PlayList.DoesNotExist:
                             return Response({"detail": "찾을 수 없습니다."}, status=status.HTTP_404_NOT_FOUND)
